+++ src/roq/logging/flags/flags.hpp
@@ -18,6 +18,7 @@ struct Flags final {
   static uint32_t log_max_files();
   static bool log_rotate_on_open();
   static std::string_view color();
+  static uint32_t log_verbosity();
 };
 
 }  // namespace flags